:recycle: refactor: GameController의 게임 로직을 Rcing 클래스로 분리
