+++ src/main/java/racingcar/controller/GameController.java
@@ -3,6 +3,7 @@ package racingcar.controller;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
 import racingcar.model.Car;
+import racingcar.model.Racing;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
 
@@ -11,37 +12,8 @@ public class GameController {
         List<Car> cars = InputView.getCarNames();
         int numberOfAttempts = InputView.getNumberOfAttempts();
 
-        List<String> winners = startRace(cars, numberOfAttempts);
+        List<String> winners = Racing.startRace(cars, numberOfAttempts);
 
         OutputView.displayWinners(winners);
-    }
-
-    private static List<String> startRace(List<Car> cars, int numberOfAttempts) {
-        OutputView.displayResult();
-        for (int attempt = 0; attempt < numberOfAttempts; attempt++) {
-            makeCarMovements(cars);
-            OutputView.displayCarMovements(cars);
-        }
-        return announceWinners(cars);
-    }
-
-    private static void makeCarMovements(List<Car> cars) {
-        for (Car car : cars) {
-            if (Randoms.pickNumberInRange(0, 9) >= 4) {
-                car.increasePosition();
-            }
-        }
-    }
-
-    private static List<String> announceWinners(List<Car> cars) {
-        int maxPosition = cars.stream()
-                .mapToInt(Car::getPosition)
-                .max()
-                .orElse(0);
-
-        return cars.stream()
-                .filter(car -> car.getPosition() == maxPosition)
-                .map(Car::getName)
-                .toList();
     }
 }
